+++ src/Sentry/Internal/Hub.cs
@@ -69,6 +69,8 @@ internal class Hub : IHub, IDisposable
             PushScope();
         }
 
+        Logger = SentryStructuredLogger.Create(this, options, _clock);
+
 #if MEMORY_DUMP_SUPPORTED
         if (options.HeapDumpOptions is not null)
         {
@@ -829,6 +831,9 @@ internal class Hub : IHub, IDisposable
         _memoryMonitor?.Dispose();
 #endif
 
+        Logger.Flush();
+        Logger.Dispose();
+
         try
         {
             CurrentClient.FlushAsync(_options.ShutdownTimeout).ConfigureAwait(false).GetAwaiter().GetResult();
@@ -853,4 +858,6 @@ internal class Hub : IHub, IDisposable
     }
 
     public SentryId LastEventId => CurrentScope.LastEventId;
+
+    public SentryStructuredLogger Logger { get; }
 }